Reworked Note Handling
NoteArrow's handle their own auto miss and registering to be hit
Arrows are no pooled and spawned on demand instead of looping endlessly
CM purely handles spawning and pooling arrows
CD manages NoteArrowData for knowing when to spawn notes, and managing calculating timing
BattleDirector purely handles battle related things
Will work on documentation
Most floats are changed to doubles for precision security

*Also changed rewardSelect to use Stages instead of string*

+++ Globals/TimeKeeper.cs
@@ -7,8 +7,8 @@ using Godot;
 public partial class TimeKeeper : Node
 {
     public static double CurrentTime = 0;
-    public static float ChartLength;
-    public static float LoopLength;
+    public static double ChartLength;
+    public static double LoopLength;
     public static float Bpm;
     public static Beat LastBeat { get; set; }
     public static double BeatsPerLoop;
